refactor: move EventType and EventSource enums to pojos/enums package (#29)
Enums belong in pojos/enums, not under dto. Removed the dto/kafka directory
and updated all imports accordingly.
